% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAzureADToken.R
\name{getAzureADToken}
\alias{getAzureADToken}
\title{AzureAD Token の取得}
\usage{
getAzureADToken(TENANT_ID, CLIENT_ID, CLIENT_SECRET)
}
\arguments{
\item{TENANT_ID}{character}

\item{CLIENT_ID}{character}

\item{CLIENT_SECRET}{character}
}
\value{
TOKEN character Azure AD Token
}
\description{
AzureAD Token の取得
}
