% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAzureCSPPartnerCenterToken.R
\name{getAzureCSPPartnerCenterToken}
\alias{getAzureCSPPartnerCenterToken}
\title{Azure CSP PartnetCenter Token の取得}
\usage{
getAzureCSPPartnerCenterToken(DOMAIN, CLIENT_ID, CLIENT_SECRET)
}
\arguments{
\item{DOMAIN}{character}

\item{CLIENT_ID}{character}

\item{CLIENT_SECRET}{character}
}
\value{
TOKEN character CSP PartnetCenter Token
}
\description{
Azure CSP PartnetCenter Token の取得
}
