% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCSPPriceTable.R
\name{getCSPPriceTable}
\alias{getCSPPriceTable}
\title{CSP 価格表の取得}
\usage{
getCSPPriceTable(PC_TOKEN)
}
\arguments{
\item{PC_TOKEN}{character Azure CSP PartnetCenter Token}
}
\value{
ratecard data.frame
}
\description{
CSP 価格表の取得
}
