% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPriceTable.R
\name{getPriceTable}
\alias{getPriceTable}
\title{定価表の取得}
\usage{
getPriceTable(SUBSCRIPTION_ID, ADTOKEN)
}
\arguments{
\item{SUBSCRIPTION_ID}{character サブスクリプションID}

\item{ADTOKEN}{character AD Token}
}
\value{
pricelist data.frame
}
\description{
定価表の取得
}
